AzureFunctions/negotiate/__init__.py
import logging

import azure.functions as func


def main(req: func.HttpRequest, connectionInfo: str) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')
    return func.HttpResponse(
        connectionInfo,
        status_code=200,
        headers={
            'Content-type': 'application/json'
        }
    )
