AzureFunctions/requirements.txt
azure-common==1.1.27
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.7.2
azure-nspkg==3.0.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.1
cryptography==3.4.7
idna==3.2
pycparser==2.20
python-dateutil==2.8.2
python-http-client==3.3.2
requests==2.26.0
sendgrid==6.7.1
six==1.16.0
starkbank-ecdsa==1.1.1
urllib3==1.26.6
